requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
decorator==5.1.1
ffmpeg-python==0.2.0
filelock==3.17.0
fsspec==2024.12.0
future==1.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.5
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
pillow==10.4.0
proglog==0.1.10
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sympy==1.13.1
tiktoken==0.8.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
streamlit~=1.42.2
librosa~=0.10.2.post1